+++ processor/src/test/resources/input/com/example/inject/InjectTrueComponent.java
@@ -1,6 +1,7 @@
 package com.example.inject;
 
 import javax.inject.Inject;
+import react4j.annotations.Feature;
 import react4j.annotations.ReactComponent;
 import react4j.core.BaseContext;
 import react4j.core.BaseProps;
@@ -8,7 +9,7 @@ import react4j.core.BaseState;
 import react4j.core.Component;
 import react4j.core.ReactNode;
 
-@ReactComponent( inject = ReactComponent.Feature.ENABLE )
+@ReactComponent( inject = Feature.ENABLE )
 abstract class InjectTrueComponent
   extends Component<BaseProps, BaseState, BaseContext>
 {